refactor: split hybridSearch.ts into three focused modules
- queryDecomposition.ts: LLM query decomposition (70 lines)
- hybridSearch.ts: keyword search + RRF + keyword reranking (145 lines)
- searchOrchestration.ts: MaxSim scoring + search_by_note/query (155 lines)

No behaviour change. Each module has one concern. Updated all importers.
