+++ src/notes/hybridSearch.ts
@@ -1,75 +1,13 @@
-import joplin from 'api';
-import { BlockEmbedding, NoteEmbedding, find_nearest_notes, group_by_notes, corpusCaches, userDataStore } from './embeddings';
-import { read_user_data_embeddings } from './userDataReader';
-import { TextEmbeddingModel, TextGenerationModel } from '../models/models';
-import { JarvisSettings } from '../ux/settings';
-import { clearApiResponse, with_timeout } from '../utils';
-import { calc_similarity } from './embeddingHelpers';
-import { quantize_vector_to_q8 } from './q8';
-import { SimpleCorpusCache } from './embeddingCache';
-import { getLogger } from '../utils/logger';
-
-const log = getLogger();
-
 /**
- * Decompose a user query into focused sub-queries for hybrid search.
- * Each sub-query has a semantic component (for embedding search) and
- * keyword terms (for Joplin search API).
+ * Keyword search and Reciprocal Rank Fusion (RRF) for hybrid retrieval.
  *
- * @param query - the user's question
- * @param model_gen - text generation model for LLM call
- * @returns array of sub-queries, or null on failure/timeout
+ * Provides keyword search via Joplin's search API, RRF merging of
+ * semantic + keyword results, and keyword reranking of scored blocks.
  */
-export async function decompose_query(
-  query: string,
-  model_gen: TextGenerationModel,
-): Promise<{semantic: string, keywords: string[]}[] | null> {
-  const prompt = `Decompose this question into 1-3 focused search sub-queries.
-For each, output on a separate line:
-SEARCH: <semantic query> | KEYWORDS: <terms or NONE>
-
-Rules:
-- Use 1 sub-query when the question targets a single topic, entity, or time period.
-- Use 2-3 only for genuinely multi-faceted questions (comparisons, multiple entities).
-- Use "quoted phrases" for compound terms. Drop evaluative words from keywords.
-- Combine co-occurring entities in a single keyword term.
-
-Question: ${query}`;
-
-  try {
-    // use temperature 0 for deterministic decomposition
-    const saved_temperature = model_gen.temperature;
-    model_gen.temperature = 0;
-    let response: string;
-    try {
-      response = await with_timeout(10_000, model_gen.complete(prompt));
-    } finally {
-      model_gen.temperature = saved_temperature;
-    }
-    if (!response) { return null; }
-
-    const results: {semantic: string, keywords: string[]}[] = [];
-    for (const line of response.split('\n')) {
-      const match = line.match(/SEARCH:\s*(.+?)\s*\|\s*KEYWORDS:\s*(.+)/i);
-      if (!match) { continue; }
-
-      const semantic = match[1].trim();
-      if (!semantic) { continue; }
-
-      const kw_raw = match[2].trim();
-      const keywords = kw_raw.toLowerCase() === 'none'
-        ? []
-        : kw_raw.split(',').map(k => k.trim()).filter(k => k.length > 0);
-
-      results.push({ semantic, keywords });
-    }
-
-    return results.length > 0 ? results : null;
-  } catch (error) {
-    log.info(`[Hybrid] decomposition failed: ${error.message || error}`);
-    return null;
-  }
-}
+import joplin from 'api';
+import { BlockEmbedding } from './embeddings';
+import { JarvisSettings } from '../ux/settings';
+import { clearApiResponse } from '../utils';
 
 /**
  * Return all semantic pool blocks from keyword-matched notes.
@@ -204,143 +142,4 @@ export async function keyword_rerank(
   const merged_keys = new Set(merged.map(b => `${b.id}:${b.line}`));
   const tail = scored.filter(b => !merged_keys.has(`${b.id}:${b.line}`));
   return [...merged, ...tail];
-}
-
-/**
- * MaxSim scoring: for each pool block, compute max cosine similarity
- * across all query embeddings (ColBERT-style late interaction).
- * Sets block.similarity in place.
- *
- * @param query_embeddings - query vectors (one per chunk/turn/sub-query)
- * @param pool - blocks to score (modified in place: .similarity set)
- * @param exclude_id - note ID to exclude from scoring (current note)
- */
-export function maxsim_score(
-  query_embeddings: Float32Array[],
-  pool: BlockEmbedding[],
-  exclude_id: string,
-): void {
-  for (const block of pool) {
-    if (block.id === exclude_id) { continue; }
-    let max_sim = 0;
-    for (const query_emb of query_embeddings) {
-      const sim = calc_similarity(block.embedding, query_emb);
-      if (sim > max_sim) { max_sim = sim; }
-    }
-    block.similarity = max_sim;
-  }
-}
-
-/**
- * MaxSim search across cache or legacy pool.
- * For each pool block, compute max cosine similarity across query embeddings.
- * Uses Q8 cache when available, falls back to Float32 on in-memory pool.
- *
- * @param query_embeddings - query vectors (one per chunk/turn/sub-query)
- * @param pool - legacy in-memory pool (may be empty in userData mode)
- * @param cache - corpus cache (may be undefined or unbuilt)
- * @param exclude_id - note ID to exclude (current note)
- * @param settings - for min_similarity, min_length, notes_max_hits
- * @returns scored blocks sorted by similarity descending, or empty if no pool available
- */
-export function maxsim_search(
-  query_embeddings: Float32Array[],
-  pool: BlockEmbedding[],
-  cache: SimpleCorpusCache | undefined,
-  exclude_id: string,
-  settings: JarvisSettings,
-): BlockEmbedding[] {
-  if (cache?.isBuilt()) {
-    // userData: per-query cache search, keep max similarity per block
-    const block_scores = new Map<string, BlockEmbedding>();
-    for (const emb of query_embeddings) {
-      const q8 = quantize_vector_to_q8(emb);
-      const results = cache.search(q8, settings.notes_max_hits * 4, settings.notes_min_similarity);
-      for (const r of results) {
-        if (r.id === exclude_id) { continue; }
-        const key = `${r.id}:${r.line}`;
-        const existing = block_scores.get(key);
-        if (!existing || r.similarity > existing.similarity) {
-          block_scores.set(key, r);
-        }
-      }
-    }
-    const scored = [...block_scores.values()];
-    scored.sort((a, b) => b.similarity - a.similarity);
-    return scored;
-  }
-
-  if (pool.length > 0) {
-    // legacy: Float32 MaxSim on in-memory pool
-    maxsim_score(query_embeddings, pool, exclude_id);
-    const filtered = pool.filter(b =>
-      b.id !== exclude_id &&
-      b.similarity >= settings.notes_min_similarity &&
-      b.length >= settings.notes_min_length);
-    filtered.sort((a, b) => b.similarity - a.similarity);
-    return filtered;
-  }
-
-  return [];  // cold start: no pool available
-}
-
-/**
- * Multi-chunk search for a note: load chunks, MaxSim score, keyword rerank, group.
- * Used by the related notes panel and the API when a noteId is provided.
- *
- * @returns grouped NoteEmbedding[], or null if chunks unavailable or multi-chunk disabled
- */
-export async function search_by_note(
-  noteId: string,
-  noteTitle: string,
-  model: TextEmbeddingModel,
-  settings: JarvisSettings,
-): Promise<NoteEmbedding[] | null> {
-  if (!settings.notes_multi_chunk_search) { return null; }
-
-  // load note's chunk embeddings (legacy pool or userData)
-  let query_chunks: BlockEmbedding[] = model.embeddings.filter(b => b.id === noteId);
-  if (query_chunks.length === 0 && settings.notes_db_in_user_data) {
-    const loaded = await read_user_data_embeddings({
-      store: userDataStore, modelId: model.id, noteIds: [noteId],
-    });
-    if (loaded.length > 0) { query_chunks = loaded[0].blocks; }
-  }
-  if (query_chunks.length === 0) { return null; }
-
-  const cache = corpusCaches.get(model.id);
-  const scored = maxsim_search(
-    query_chunks.map(c => c.embedding), model.embeddings, cache, noteId, settings);
-  if (scored.length === 0) { return null; }
-
-  const reranked = noteTitle
-    ? await keyword_rerank(scored, [noteTitle], settings)
-    : scored;
-  return group_by_notes(reranked, settings);
-}
-
-/**
- * Text query search: flat semantic search, keyword rerank, group by note.
- * Used by the panel search box and the API when a text query is provided.
- *
- * @returns grouped NoteEmbedding[]
- */
-export async function search_by_query(
-  query: string,
-  excludeId: string,
-  model: TextEmbeddingModel,
-  settings: JarvisSettings,
-  panel?: string,
-  isUpdateInProgress?: boolean,
-): Promise<NoteEmbedding[]> {
-  const flat = await find_nearest_notes(
-    model.embeddings, excludeId, 1, '', query,
-    model, settings, false, panel, isUpdateInProgress ?? false,
-  );
-  if (flat.length === 0 || flat[0].embeddings.length === 0) { return flat; }
-
-  const reranked = await keyword_rerank(flat[0].embeddings, [query], settings);
-  return reranked.length > 0
-    ? group_by_notes(reranked, settings)
-    : flat;
 }
